Major Updates

+++ database/migrations/2024_08_03_091924_creates_assessment_questions_table.php
@@ -9,6 +9,9 @@ return new class extends Migration
    
     public function up()
     {
+        if( Schema::hasTable('assessment_questions') ) return ;
+
+        
         Schema::create('assessment_questions', function(Blueprint $table){
             $table->ulid('uuid')->unique()->index();
             $table->foreignUlid('assessment_id')->nullable()->constrained( table: 'assessments', column: 'uuid' )->onDelete('set null');
